scripts/requirements.txt
et-xmlfile==1.1.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
